+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Introducing layouts in `PageLayout`. See the [`PageLayout`](http://mapbox.github
   - The added layouts: `page`, `accordion`, `example`, `full`, `exampleIndex`.
   - 🚨 The following props are deprecated and are no longer configurable: `sidebarTitle`, `sidebarContent`, `sidebarContentStickyTop`, `sidebarContentStickyTopNarrow`, `sidebarStackedOnNarrowScreens`, `sideBarColSize`, `interactiveClass`.
   - `sidebarTheme` is now defined from the `frontMatter` object.
+  - Component now includes `TopbarSticker/Topbar`, `Search`, and `Feedback` components.
 - 🚨 The `user` props is deprecated in `Feedback`. The component will now fetch this data and you can safely remove the `user` prop from this component. [#324](https://github.com/mapbox/dr-ui/pull/324)
 - Update dependencies. [#326](https://github.com/mapbox/dr-ui/pull/326)